Promote MOSS codec runtime to framework
