+++ include/engine/models/moss/moss_tts_local/generator.h
@@ -3,6 +3,7 @@
 #include "engine/models/moss/moss_tts_local/assets.h"
 #include "engine/models/moss/moss_tts_local/backbone.h"
 #include "engine/models/moss/moss_tts_local/depth_transformer.h"
+#include "engine/framework/modules/multi_codebook_embedding.h"
 #include "engine/models/moss/shared/token_rows.h"
 #include "engine/framework/sampling/torch_random.h"
 
@@ -65,7 +66,7 @@ private:
     const MossDepthTransformer & depth_;
     int64_t hidden_size_ = 0;
     int64_t num_codebooks_ = 0;
-    std::unique_ptr<moss::AudioCodebookEmbeddings> audio_codebooks_;
+    std::unique_ptr<engine::modules::MultiCodebookEmbedding> audio_codebooks_;
     std::vector<float> local_text_head_;                // [2 * hidden]
     engine::sampling::TorchCudaSamplingPolicy sampling_policy_;
     struct ProjectionRuntime;
